pyproject.toml
# ═══════════════════════════════════════════════════════════════════════════
# PYPROJECT.TOML — iAngel Alpha Backend
# Configuration Python stricte avec uv + ruff + mypy
# ═══════════════════════════════════════════════════════════════════════════

[project]
name = "iangel-alpha"
version = "0.1.0"
description = "iAngel - Ange-gardien numérique des techno-vulnérables"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [{ name = "Fil", email = "fil@dataml.ca" }]

dependencies = [
    "fastapi>=0.109.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "uvicorn[standard]>=0.25.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "anthropic>=0.18.0",
    "structlog>=24.0",
    "tenacity>=9.1.2",
    "sqlalchemy[asyncio]>=2.0.45",
    "asyncpg>=0.31.0",
    "aiosqlite>=0.22.1",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.8",
    "ruff>=0.3",
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.23",
]

# === UV ===
[tool.uv]
dev-dependencies = [
    "mypy>=1.8",
    "ruff>=0.3",
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.23",
]

# ═══════════════════════════════════════════════════════════════════════════
# RUFF — Linter et Formatter ultrarapide
# ═══════════════════════════════════════════════════════════════════════════

[tool.ruff]
target-version = "py311"
line-length = 88
fix = true

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ANN",    # flake8-annotations
    "S",      # flake8-bandit (security)
    "BLE",    # flake8-blind-except
    "RUF",    # Ruff-specific rules
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "PTH",    # flake8-use-pathlib
]

ignore = [
    "E501",   # line too long (géré par formatter)
    "ANN101", # missing type annotation for self
    "ANN102", # missing type annotation for cls
    "S101",   # assert (ok dans code app pour invariants)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101",    # assert allowed in tests
    "ARG",     # unused args ok in fixtures
    "ANN",     # annotations relaxées dans tests
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

# ═══════════════════════════════════════════════════════════════════════════
# MYPY — Type checking strict
# ═══════════════════════════════════════════════════════════════════════════

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
show_column_numbers = true
pretty = true

plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false

# ═══════════════════════════════════════════════════════════════════════════
# PYTEST — Configuration des tests
# ═══════════════════════════════════════════════════════════════════════════

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "-ra",
    "-q",
    "--strict-markers",
    "--strict-config",
    "-vv",
]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
]
markers = [
    "unit: unit tests",
    "integration: integration tests",
    "slow: marks tests as slow",
]

[tool.coverage.run]
source = ["app"]
branch = true
omit = ["tests/*", "*/__pycache__/*"]

[tool.coverage.report]
fail_under = 80
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
